backend/requirements.txt
asgiref==3.4.1
astroid==2.7.2
autopep8==1.5.7
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.2.0
pycodestyle==2.7.0
pylint==2.10.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
gunicorn==20.1.0